Share target excess redistribution helper
